Moved ScrollToTop into s1 folder and updated imports in App.jsx
