+++ src/App.jsx
@@ -1,7 +1,6 @@
 import "./App.css";
 import { BrowserRouter } from "react-router-dom";
-
-import { ScrollToTop } from "./ScrollToTop";
+import { ScrollToTop } from "./s1/ScrollToTop";
 
 export const App = () => {
   return (
